rename packages and properties
