+++ integration/process-engine-api/engine-client/src/main/kotlin/io/holunda/polyflow/client/processengineapi/EngineClientAutoConfiguration.kt
@@ -1,4 +1,4 @@
-package io.holunda.polyflow.client.camunda
+package io.holunda.polyflow.client.processengineapi
 
 import io.holunda.polyflow.spring.ApplicationNameBeanPostProcessor
