+++ integration/process-engine-api/engine-client/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-io.holunda.polyflow.client.camunda.EngineClientAutoConfiguration
+io.holunda.polyflow.client.processengineapi.EngineClientAutoConfiguration
