+++ integration/process-engine-api/engine-client/src/test/kotlin/io/holunda/polyflow/client/processengineapi/CamundaEngineClientPropertiesTestApplication.kt
@@ -1,4 +1,4 @@
-package io.holunda.polyflow.client.camunda
+package io.holunda.polyflow.client.processengineapi
 
 import io.holunda.polyflow.spring.ApplicationNameBeanPostProcessor
 import org.springframework.boot.autoconfigure.SpringBootApplication
